requirements.txt
selenium==4.18.1
pytest==8.0.0
pytest-html==4.1.1
webdriver-manager==4.0.1
allure-pytest==2.13.2
allure-python-commons~=2.13.2


